cognigate-api/pyproject.toml
[project]
name = "cognigate-api"
version = "0.2.0"
description = "Cognigate Engine - AI Agent Governance API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Vorion", email = "dev@vorion.org"}
]
keywords = ["ai", "governance", "agents", "basis", "trust"]

dependencies = [
    "fastapi>=0.131.0",
    "uvicorn[standard]>=0.41.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0.1",
    "python-multipart",
    # Database (PostgreSQL/Neon)
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    # Cryptography for proof signatures
    "cryptography>=42.0.0",
    # Redis for caching and distributed state
    "redis>=5.0.0",
    # AI Providers for Critic Pattern
    "anthropic>=0.18.0",
    "openai>=1.12.0",
    "google-generativeai>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
